Module task_graph::scheduler

Structs

SequentialScheduler
Scheduler that immediatly runs the function in the same thread

SpawningScheduler
Scheduler that spawns a new OS thread for each scheduled function

TestScheduler
Scheduler that queues all scheduled functions to allow for inspection and later running them in a controlled manner, e.g. in tests

Traits

Scheduler
Trait used for generic scheduling of work