Struct task_graph::task::Task

pub struct Task<A: Send + 'static> {
    // some fields omitted
}

A computation that eventually will produce a value that other tasks can depend on.

A graph of dependent tasks can be created by chaining together several tasks with then and join. Parallelism is achived when multiple tasks depend on the same cloned task. Cloning a task will clone the produced value, but not run the function twice. If the produced value is not Clone, the function arc can be used to wrap the value in an Arc (but will then require the original value to be Sync). There is virtually no locking involved; tasks will simply not be scheduled until all dependent tasks are done. A task can be converted into a Future in order to allow blocking for a task to finish and to retrieve its result.

The result from the computation will need to be Send. Task implements Send itself and can thus be used inside and returned from other tasks (effectively forming a monad).

Methods

impl<A> Task<A> where A: Send + 'static

fn from_value(val: A) -> Task<A>

Create a finished task from a value

fn from_fn<F, S>(sched: &S, fun: F) -> Task<A> where F: Send + FnOnce() -> A + 'static, S: Scheduler + 'static

Create a task from a function that will be scheduled to run in sched.

fn never() -> Task<A>

Create a task that never finishes

Tasks made dependent on this task will never be scheduled or run.

fn then<B, F, S>(self, sched: &S, fun: F) -> Task<B> where F: Send + FnOnce(A) -> B + 'static, S: Scheduler + 'static, B: Send + 'static

Create a task that is a continuation from this task

This continuation will run in parallel with other continuations created from clones of the same task. The function fun will be scheduled on sched as soon as the current task that the continuation was created from is finished.

fn then_forget<F, S>(self, sched: &S, fun: F) where F: Send + FnOnce(A) + 'static, S: Scheduler + 'static

Run a function as a continuation from this task, but don't return it in a task

This continuation will run in parallel with other continuations created from clones of the same task. The function fun will be scheduled on sched as soon as the current task that the continuation was created from is finished.

fn join<B>(self, other: Task<B>) -> Task<(A, B)> where A: 'static, B: Send + 'static

Create a task that contains the result of this and another task.

Continuations on this task will run in parallel with other continuations created from clones of these tasks. The returned task will be done as soon as both the current task and the other task are finished.

fn into_future(self) -> Future<A> where A: 'static

Turn this task into a future, from which the result can be retrieved

fn is_done(&self) -> bool

Returns if this task has finished

Mostly only useful for testing, or together with waiting for a future returned from into_future

impl<A> Task<A> where A: Send + Sync + 'static

fn arc(self) -> Task<Arc<A>>

impl<A> Task<Task<A>> where A: Send + 'static

fn unwrap(self) -> Task<A>

Retrieve the value from the inner task

This is removing one layer of Task without blocking, i.e. the resulting task will not run until the inner task is finished.

impl<A> Task<Vec<A>> where A: Send + 'static

fn for_each<B: Send, F, S>(self, sched: &S, fun: F) -> Task<Vec<B>> where F: Send + Sync + Fn(A) -> B + 'static, S: Scheduler + 'static