Trait task_graph::task::Join

pub trait Join<A> {
    fn join_all(self) -> Task<Vec<A>>;
    fn join_any(self) -> Task<A>;
}

Trait for joining tasks in Vec<Task<A>>

Required Methods

fn join_all(self) -> Task<Vec<A>>

Create a task that contains all the results of the input tasks

Continuations on the returned task will run in parallel with other continuations created from these tasks. The returned task will be done as soon as all the tasks in the Vec are finished. The returned task will contain all the input tasks values, but not necessarily in the same order. This function works even with an empty Vec of input tasks, in which case it is done immediatly (containing an empty Vec).

fn join_any(self) -> Task<A>

Create a task that contains one of the results of the input tasks

Continuations on the returned task will run in parallel with other continuations created from these tasks. If at least one of the returned tasks are already done, then the returned task will be done and contain the value of that task. Otherwise the returned task will be done as soon as the first input task is done and containe that tasks's value.

Running join_any on an empty Vec, will result in a task that never finishes (just as Task::never()).